Refactor: global-exception 안에 디렉토리 구조 리팩토링해라 - #34
global-exception 안에 디렉토리 구조 리팩토링함.

리팩토링한 내용:
- exceptions 디렉토리 생성 후 안에 예외 관리함
