+++ src/main/java/org/example/youngnam/domain/image/service/ImageServiceImpl.java
@@ -10,9 +10,9 @@ import org.example.youngnam.domain.image.dto.response.ImageResponseDTO;
 import org.example.youngnam.domain.image.entity.Image;
 import org.example.youngnam.domain.image.mapper.ImageMapper;
 import org.example.youngnam.domain.image.repository.ImageRepository;
-import org.example.youngnam.global.exception.EntityNotFoundException;
+import org.example.youngnam.global.exception.exceptions.EntityNotFoundException;
 import org.example.youngnam.global.exception.ErrorCode;
-import org.example.youngnam.global.exception.UnauthorizedException;
+import org.example.youngnam.global.exception.exceptions.UnauthorizedException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
