+++ src/main/java/org/example/youngnam/global/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package org.example.youngnam.global.exception;
 
+import org.example.youngnam.global.exception.exceptions.YoungnamException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
